requirements.txt
# CommercePulse — Python dependencies
# Install: pip install -r requirements.txt

# ── Core data stack ───────────────────────────────────────────
duckdb>=0.10.0
pandas>=2.0.0
pyarrow>=14.0.0          # fast Parquet I/O, DuckDB integration
pyyaml>=6.0.1

# ── API layer ─────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Dashboard ─────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.19.0

# ── dbt transform ─────────────────────────────────────────────
dbt-core>=1.8.0
dbt-duckdb>=1.8.0

# ── Testing ───────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.26.0            # async test client for FastAPI

# ── Dev tooling ───────────────────────────────────────────────
ruff>=0.3.0              # linting + formatting
mypy>=1.8.0              # type checking
python-dotenv>=1.0.0     # .env file support
